requirements.txt
requests
pycryptodome
beautifulsoup4